+++ Elements/src/Geometry/Polygon.cs
@@ -311,7 +311,7 @@ namespace Elements.Geometry
 
             if (sort)
             {
-                result.Sort(new DotComparer(d));
+                result.Sort(new DirectionComparer(d));
             }
 
             return result.Count > 0;
@@ -372,7 +372,7 @@ namespace Elements.Geometry
             if (sort)
             {
                 // Order the intersections along the direction.
-                results.Sort(new DotComparer(d));
+                results.Sort(new DirectionComparer(d));
             }
 
             return results.Count > 0;
@@ -928,7 +928,7 @@ namespace Elements.Geometry
                 var d = this.Normal().Cross(planes[i].Normal).Unitized();
                 if (results[i].Count > 0)
                 {
-                    results[i].Sort(new DotComparer(d));
+                    results[i].Sort(new DirectionComparer(d));
                 }
             }